// backend/src/prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OperationType = {
  MINT: 'MINT',
  BURN: 'BURN',
  FREEZE: 'FREEZE'
} as const

export type OperationType = (typeof OperationType)[keyof typeof OperationType]


export const MultisigKind = {
  MINT_BURN: 'MINT_BURN',
  ADMIN: 'ADMIN'
} as const

export type MultisigKind = (typeof MultisigKind)[keyof typeof MultisigKind]